+++ lib/src/logging/loggers.dart
@@ -1,4 +1,0 @@
-const String buildRunner = 'BuildRunner';
-const String parser = 'Parser';
-const String proxy = 'Proxy';
-const String webSocket = 'WebSocket';