+++ lib/src/server/proxy.dart
@@ -1,39 +1,154 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:logging/logging.dart';
+import 'package:args/args.dart';
 import 'package:mime/mime.dart';
 import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as io;
-import 'package:shelf_proxy/shelf_proxy.dart' as shelfProxy;
+import 'package:shelf_proxy/shelf_proxy.dart';
 
-import '../constants.dart';
-import '../logging/loggers.dart' as loggers;
+import '../logging.dart';
+import '../parser.dart';
+import 'build_runner.dart';
+import 'signals.dart';
+import 'websocket.dart';
 
-const _contentType = 'Content-Type';
-const _textHtml = 'text/html';
-
-/// Starts a server at [uri] proxying requests to [buildRunnerUri].
+/// A simple `shelf` server which will proxy requests to [to].
 ///
-/// *This needs some understanding of [`package: shelf`](https://pub.dartlang.org/packages/shelf).*
-Future<Null> startProxyServer(Uri uri, Uri buildRunnerUri,
-    [Pipeline pipeline = const Pipeline()]) async {
-  final handler = shelfProxy.proxyHandler(buildRunnerUri);
-  final server =
-      await io.serve(pipeline.addHandler(handler), uri.host, uri.port);
-  new Logger(loggers.proxy).info(
-      'Running Proxy server at http://${server.address.host}:${server.port}');
+/// Call [serve] to start the server at [uri].
+/// A [pipeline] can be added to customize the behavior of the server.
+///
+/// *This needs some understanding of [`package: shelf`][].*
+///
+/// [`package: shelf`]: https://pub.dartlang.org/packages/shelf
+class ProxyServer {
+  static const defaultPort = 8000;
+
+  final Uri uri;
+  final Uri to;
+  final Pipeline pipeline;
+
+  ProxyServer(this.uri, this.to, [this.pipeline = const Pipeline()]);
+
+  /// Set up a [ProxyServer] with arguments parsed by [liveReloadArgParser].
+  factory ProxyServer.fromParsed(ArgResults results, Uri to,
+      [Pipeline pipeline = const Pipeline()]) {
+    return new ProxyServer(_getUri(results), to, pipeline);
+  }
+
+  /// Starts this server at [uri] and will log a message returned from [logMessage] when this server is successfully started.
+  ///
+  /// *This method must be called only once.*
+  Future<Null> serve(String logMessage(HttpServer server)) async {
+    if (_serveHasBeenCalled)
+      throw new StateError(
+          'For each instance, `serve` must be called only once.');
+    _serveHasBeenCalled = true;
+
+    _server = await io.serve(
+        pipeline.addHandler(proxyHandler(to)), uri.host, uri.port);
+    logger.info(logMessage(_server));
+  }
+
+  /// Forces the server to close.
+  Future<dynamic> forceClose() => _server.close(force: true);
+
+  /// Extracts the proxy server uri from arguments parsed by [liveReloadArgParser].
+  static Uri _getUri(ArgResults results) {
+    return new Uri(
+        scheme: 'http',
+        host: results[CliOption.hostName] as String,
+        port:
+            int.parse(results[CliOption.proxyPort] as String, onError: (input) {
+          logger.warning(
+              '${RecordPrefix.warning} The port number `$input` must be an integer. Default proxy port `$defaultPort` is used instead.\n');
+          return defaultPort;
+        }));
+  }
+
+  /// Keeps ref to the server just for closing.
+  HttpServer _server;
+
+  /// Indicates if [serve] has been called.
+  bool _serveHasBeenCalled = false;
 }
 
-Pipeline liveReloadPipeline(Uri webSocketUri) => const Pipeline()
-    .addMiddleware(injectJavaScript(reloadOn(webSocketUri, reloadSignal)));
+/// A livereload proxy server.
+///
+/// This server proxies requests to `build_runner serve` and injects a script that will reload the browser on [reloadSignal]. (See [injectJavaScript] and [reloadOn].)
+/// If this serves a single page application, it will rewrite any 404 response whose path doesn't have a MIME type to the root path of `build_runner serve`. (See [rewriteTo] and [shouldBeRewritten].)
+///
+/// If this heuristic doesn't work well for you, please file an issue. Furthermore, PRs are welcome.
+///
+/// Call [serve] to start the server at [uri].
+/// A [pipeline] can be added to customize the behavior of the server.
+///
+/// This should be used together with [LiveReloadWebSocketServer].
+///
+/// *This needs some understanding of [`package: shelf`][].*
+///
+/// [`package: shelf`]: https://pub.dartlang.org/packages/shelf
+class LiveReloadProxyServer extends ProxyServer {
+  /// Returns `true` for any 404 response whose path doesn't have a MIME type.
+  static bool shouldBeRewritten(Request request, Response response) {
+    return response.statusCode == 404 &&
+        request.url.pathSegments.isNotEmpty &&
+        lookupMimeType(request.url.pathSegments.last) == null &&
+        // .map doesn't have a MIME type, a workaround is needed
+        !request.url.path.endsWith('.map');
+  }
 
-Pipeline liveReloadSpaPipeline(Uri webSocketUri) =>
-    liveReloadPipeline(webSocketUri).addMiddleware(rewriteTo('/'));
+  /// A future which will resolve when `build_runner` starts serving.
+  ///
+  /// This expects [BuildRunnerServeProcess.served].
+  final Future<Null> buildRunnerServed;
+  final Uri buildRunnerUri;
+  final Uri webSocketUri;
+
+  LiveReloadProxyServer(Uri uri, this.buildRunnerUri, this.webSocketUri,
+      this.buildRunnerServed, bool isSinglePageApplication,
+      [Pipeline pipeline = const Pipeline()])
+      : super(
+            uri,
+            buildRunnerUri,
+            isSinglePageApplication
+                ? pipeline
+                    .addMiddleware(
+                        injectJavaScript(reloadOn(webSocketUri, reloadSignal)))
+                    .addMiddleware(rewriteTo('/', shouldBeRewritten))
+                : pipeline.addMiddleware(
+                    injectJavaScript(reloadOn(webSocketUri, reloadSignal))));
+
+  /// Set up a [LiveReloadProxyServer] with arguments parsed by [liveReloadArgParser].
+  factory LiveReloadProxyServer.fromParsed(
+      ArgResults results,
+      BuildRunnerServeProcess buildRunner,
+      LiveReloadWebSocketServer webSocketServer,
+      [Pipeline pipeline = const Pipeline()]) {
+    return new LiveReloadProxyServer(
+        ProxyServer._getUri(results),
+        buildRunner.uri,
+        webSocketServer.uri,
+        buildRunner.served,
+        results[CliOption.singlePageApplication] == true,
+        pipeline);
+  }
+
+  @override
+  Future<Null> serve([logMessage = _defaultMessage]) async {
+    await buildRunnerServed;
+    super.serve(logMessage);
+  }
+
+  static String _defaultMessage(HttpServer server) =>
+      '\nBrowse your web app at http://${server.address.host}:${server.port}\n\n';
+}
 
 /// Creates a [Middleware] that injects a [script] into every html response.
 ///
-/// *This needs some understanding of [`package: shelf`](https://pub.dartlang.org/packages/shelf).*
+/// *This needs some understanding of [`package: shelf`][].*
+///
+/// [`package: shelf`]: https://pub.dartlang.org/packages/shelf
 Middleware injectJavaScript(String script) =>
     createMiddleware(responseHandler: (response) async {
       if (response.headers[_contentType] == _textHtml) {
@@ -46,67 +161,31 @@ Middleware injectJavaScript(String script) =>
 
 /// Creates a [Middleware] that will replace the path of `requestedUri` with [absolutePath] if the `request` [shouldBeRewritten].
 ///
-/// By default, the `request` that [shouldBeRewritten] is the one whose url doesn't have a MIME type.
-/// If this heuristic doesn't work well for you, please file an issue. Furthermore, PRs are welcome.
+/// *This needs some understanding of [`package: shelf`][].*
 ///
-/// *This needs some understanding of [`package: shelf`](https://pub.dartlang.org/packages/shelf).*
+/// [`package: shelf`]: https://pub.dartlang.org/packages/shelf
 Middleware rewriteTo(String absolutePath,
-        [bool shouldBeRewritten(Request request) = _hasNoMimeType]) =>
-    (innerHandler) => (request) => shouldBeRewritten(request)
-        ? innerHandler(new Request(
-            request.method, request.requestedUri.replace(path: absolutePath),
-            body: request.read(),
-            context: request.context,
-            encoding: request.encoding,
-            headers: request.headers,
-            protocolVersion: request.protocolVersion))
-        : innerHandler(request);
+        bool shouldBeRewritten(Request request, Response response)) =>
+    (innerHandler) => (request) async {
+          var requestBody = await request.readAsString(request.encoding);
+          var response = await innerHandler(request.change(body: requestBody));
+          return shouldBeRewritten(request, response)
+              ? await innerHandler(new Request(request.method,
+                  request.requestedUri.replace(path: absolutePath),
+                  body: requestBody,
+                  context: request.context,
+                  encoding: request.encoding,
+                  headers: request.headers,
+                  protocolVersion: request.protocolVersion))
+              : response;
+        };
 
-/// Generates a JavaScript which will reload the browser on receving a [message] from [webSocketUri].
-String reloadOn(Uri webSocketUri, String message) =>
-    "new WebSocket('$webSocketUri').onmessage=function(e){if(e.data==='$message')window.location.reload()};";
+/// Generates a JavaScript which will reload the browser on receiving a [message] from [webSocketUri].
+String reloadOn(Uri webSocketUri, String message) => "(function(){"
+    "var a=new WebSocket('$webSocketUri');"
+    "a.onmessage=function(e){if(e.data==='$message')window.location.reload()};"
+    "window.onunload=function(e){a.send('$disconnectSignal')};"
+    "})();";
 
-/// Checks if the [request].url doesn't have a MIME type.
-bool _hasNoMimeType(Request request) {
-  return request.url.pathSegments.isNotEmpty &&
-      lookupMimeType(request.url.pathSegments.last) == null &&
-      // .map doesn't have a MIME type, a workaround is needed
-      !request.url.path.endsWith('.map');
-}
-
-/// Creates a [Middleware] that returns a `html` response with [responseBody] if the [Request.requestedUri] doesn't have a MIME type.
-///
-/// **DEPRECATED!** Use [rewriteTo] instead.
-///
-/// *This needs some understanding of [`package: shelf`](https://pub.dartlang.org/packages/shelf).*
-@deprecated
-Middleware rewriteAs(String responseBody) =>
-    createMiddleware(requestHandler: (request) {
-      return _hasNoMimeType(request)
-          ? new Response.ok(responseBody, headers: {_contentType: _textHtml})
-          : null;
-    });
-
-/// Starts a proxy server at [uri] proxying requests to [buildRunnerUri] and injects a script listening for a [reloadSignal].
-///
-/// **DEPRECATED!** Use the [liveReloadPipeline] and [liveReloadSpaPipeline] with [startProxyServer] instead.
-///
-/// *This needs some understanding of [`package: shelf`](https://pub.dartlang.org/packages/shelf).*
-@deprecated
-void startLiveReloadProxyServer(
-    Uri uri, String directory, Uri buildRunnerUri, Uri webSocketUri, bool spa,
-    [Pipeline pipeline = const Pipeline()]) {
-  final indexHtml =
-      new File.fromUri(new Uri.directory(directory).resolve('index.html'))
-          .readAsStringSync();
-  final listenForReload =
-      "new WebSocket('$webSocketUri').onmessage=function(e){if(e.data==='$reloadSignal')window.location.reload()};";
-  startProxyServer(
-      uri,
-      buildRunnerUri,
-      spa
-          ? pipeline
-              .addMiddleware(injectJavaScript(listenForReload))
-              .addMiddleware(rewriteAs(indexHtml))
-          : pipeline.addMiddleware(injectJavaScript(listenForReload)));
-}
+const _contentType = 'Content-Type';
+const _textHtml = 'text/html';